Main.js optimized in some files
