+++ api/src/controllers/user.controller.ts
@@ -32,18 +32,4 @@ export class UserController {
         }
     }
 
-    // New chat
-    static async newchat(req: Request, res: Response): Promise<void> {
-        try {
-            const data: Omit<IChat, 'id' | 'created_at'> = req.body;
-            const dataInsert = await UserModel.createChat(data);
-            console.log('[ SERVER ] New chat has been created: ' + dataInsert)
-            res.status(201).json({ data: dataInsert.user1, user2: dataInsert.user2, created_at: dataInsert.created_at })
-        } catch(error) {
-            console.log('[ SERVER ] Error createing new chat at controller: ', error)
-            throw error
-        }
-    }
-
-
 }
