+++ api/src/index.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import cors from 'cors';
 import bodyParser from 'body-parser';
 
-import router from './routes/user.routes';
+import router from './routes/router';
 
 import env from './conf/env';
